requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
pdfplumber>=0.11.0
pdf2image>=1.16.3
numpy>=1.24.0
Pillow>=10.0.0